switched code over to different activity
