+++ FiveStarStyle/app/src/main/java/com/example/fivestarstyle/MainActivity.java
@@ -1,23 +1,9 @@
 package com.example.fivestarstyle;
 
-import android.graphics.Color;
-import android.support.v4.view.ViewPager;
+import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import lecho.lib.hellocharts.model.PieChartData;
-import lecho.lib.hellocharts.model.SliceValue;
-import lecho.lib.hellocharts.view.PieChartView;
-
-import android.support.design.widget.TabLayout;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.Menu;
-import android.view.MenuItem;
 
 public class MainActivity<pieChartData> extends AppCompatActivity {
 
@@ -25,28 +11,8 @@ public class MainActivity<pieChartData> extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        createPieGraph();
-    }
 
-    public void createPieGraph() {
-        // reference pie chart view
-        PieChartView pieChartView = findViewById(R.id.chart);
-        // initialize data for the pie chart
-        List<SliceValue> pieData = new ArrayList<>();
-
-        pieData.add(new SliceValue(15, Color.parseColor("#1B4F72")).setLabel("Dresses"));
-        pieData.add(new SliceValue(29, Color.parseColor("#0d3653")).setLabel("Shoes"));
-        pieData.add(new SliceValue(18, Color.parseColor("#001E3B")).setLabel("Accessories"));
-        pieData.add(new SliceValue(18, Color.parseColor("#99B5C3")).setLabel("Tops"));
-        pieData.add(new SliceValue(20, Color.parseColor("#567d9c")).setLabel("Bottoms"));
-
-
-        // create pie chart data
-        PieChartData pieChartData = new PieChartData(pieData);
-        pieChartData.setHasLabels(true).setValueLabelTextSize(18);
-        pieChartData.setValueLabelBackgroundEnabled(false);
-        pieChartView.setPieChartData(pieChartData);
-
-        pieChartData.setSlicesSpacing(3);
+        Intent intent = new Intent(MainActivity.this, ClosetStatistics.class);
+        MainActivity.this.startActivity(intent);
     }
 }
