+++ FiveStarStyle/app/src/main/res/layout/activity_main.xml
@@ -4,76 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ClosetStatistics">
-
-    <!--Android Pie Chart View-->
-    <Button
-        android:id="@+id/signup3"
-        android:layout_width="330dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginStart="40dp"
-        android:layout_marginTop="40dp"
-        android:layout_marginEnd="41dp"
-        android:background="@color/colorPrimary"
-        android:drawableEnd="@drawable/right_arrow"
-        android:drawablePadding="30dp"
-        android:minWidth="0dp"
-        android:scaleType="center"
-        android:text="@string/btnText"
-        android:textColor="#ECF0F1"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/chart" />
-
-    <lecho.lib.hellocharts.view.PieChartView
-        android:id="@+id/chart"
-        android:layout_width="350dp"
-        android:layout_height="350dp"
-        android:layout_marginStart="55dp"
-        android:layout_marginTop="260dp"
-        android:layout_marginEnd="56dp"
-        android:layout_marginBottom="170dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="1.0" />
-
-
-    <Button
-        android:id="@+id/signup"
-        android:layout_width="330dp"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal"
-        android:layout_marginTop="16dp"
-        android:layout_marginEnd="40dp"
-        android:background="@color/colorPrimary"
-        android:drawableEnd="@drawable/right_arrow"
-        android:drawablePadding="30dp"
-        android:minWidth="0dp"
-        android:scaleType="center"
-        android:text="@string/btnText2"
-        android:textColor="#ECF0F1"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/signup3" />
-
-    <TextView
-        android:id="@+id/textView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="55dp"
-        android:layout_marginTop="10dp"
-        android:layout_marginEnd="55dp"
-        android:text="@string/titleText"
-        android:textAlignment="center"
-        android:textColor="#1B4F72"
-        android:textSize="30sp"
-        android:textStyle="bold"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.497"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
+    tools:context=".MainActivity">
 
 </android.support.constraint.ConstraintLayout>
